pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "black>=25.11.0",
    "cryptography>=46.0.3",
    "fastapi>=0.123.7",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
]
